fixes : fixed for admin:install

+++ Console/AdminInstall.php
@@ -6,6 +6,8 @@ use Illuminate\Console\Command;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Input\InputArgument;
 use Artisan;
+use Dotenv\Dotenv;
+use InvalidArgumentException;
 
 class AdminInstall extends Command
 {
@@ -41,6 +43,39 @@ class AdminInstall extends Command
     public function handle()
     {
         //
+        $force = false;
+        if (!file_exists(base_path('.env'))) {
+            $force = true;
+            \File::copy('.env.example', '.env');
+            // ask user to setup env
+            $configs = array(
+                'APP_NAME' => "Laravel Admin",
+                'APP_URL' => url("/"),
+                'DB_CONNECTION' => 'mysql',
+                'DB_HOST' => "127.0.0.1",
+                'DB_PORT' => "3306",
+                'DB_DATABASE' => "laravel_admin",
+                'DB_USERNAME' => "root",
+                'DB_PASSWORD' => "",
+            );
+
+            if (app()->environmentFilePath() != base_path('.env')) {
+                \File::copy('.env.example', '.env.example.default');
+            }
+
+            foreach ($configs as $key => $config) {
+                $input = '"'.$this->ask("Enter the value of {$key}", $config).'"';
+                $this->putEnvFile($key, $input);
+                if (app()->environmentFilePath() != base_path('.env')) {
+                    $this->putEnvFile($key, $input, app()->environmentFilePath());
+                }
+            }
+        } else {
+            // ask for force re-install?
+            $force = $this->ask('This action will refresh/re-install whole database. Do you want to continue?', false);
+        }
+        if (!$force) { return 0; }
+        Artisan::call("config:cache");
         Artisan::call('vendor:publish', ['--force' => true, '--provider' => 'Modules\Admin\Providers\AdminServiceProvider']);
         echo(Artisan::output());
         Artisan::call('migrate:fresh --seed');
@@ -49,6 +84,12 @@ class AdminInstall extends Command
         echo(Artisan::output());
         Artisan::call('module:seed');
         echo(Artisan::output());
+        Artisan::call("config:clear");
+        Artisan::call("cache:clear");
+
+        if (app()->environmentFilePath() != base_path('.env') && file_exists(base_path('.env.example.default'))) {
+            \File::move('.env.example.default', '.env.example');
+        }
     }
 
     /**
@@ -73,5 +114,21 @@ class AdminInstall extends Command
         return [
             ['example', null, InputOption::VALUE_OPTIONAL, 'An example option.', null],
         ];
+    }
+
+    public function putEnvFile($key, $value, $path=false)
+    {
+        if (!$path) {
+            $path = base_path('.env');
+            // $path = app()->environmentFilePath();
+        }
+
+        $escaped = preg_quote('='.env($key), '/');
+
+        file_put_contents($path, preg_replace(
+            "/^{$key}{$escaped}/m",
+            "{$key}={$value}",
+            file_get_contents($path)
+        ));
     }
 }
